net: netcp: ethss: move sysfs code to a separate file
For ease of maintenance and to reduce the size of ethss driver code
move the sysfs related code from netcp_ethss.c to a separate file,
netcp_ethss_sysfs.c. This requires that common definitions are moved
to a local header file netcp_ethss.h and included in both driver code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>