soc: ti: add support of Quality of Service PDSP.
The QoS PDSP is charged with the job of policing all packet flows
in the system, and verifying that neither the peripherals nor the
host CPU are overwhelmed with packets.

The key to the functionality of the QOS system is the arrangement
of packet queues. There are two distinct scheduling blocks each of
which have two setsof packet queues, the QOS ingress queues, and
the final destination queues.

There is a high priority block called the drop scheduler which
implements a model of tail drop and RED (Random Early Detect/Drop).

Next there is a medium priority block which is the QoS Scheduler.
The responsibility of QoS Scheduler is to enforce the data rate
of each configured QoS QMSS channel based on some predefined
algorithms such as Weighted Round Robin or Strict Priority

Finally there is a low priority block which enforces queue depth
on the QoS scheduler ports

This patch implements the QoS module which performs the needed
configurations in order to achieve the QoS functionality provided
by the QoS PDSP.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>
Signed-off-by: Reece R. Pollack <x0183204@ti.com>