soc: ti: qmss: add support of queue range specific push pop methods
When a range of hardware queues are monitored and
packets in those queues are proessed by a PDSP
firmware, the PDSP firmware may have a queue push
and pop methods different from the generic push
pop methods provided by the qmss module.  This patch
adds the support of such flexibility.  One application
of this is in the support of NETCP QoS.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>