soc: ti: qmss: unmap descriptors without knowing descriptor pool
Prior to this patch, it is always assumed that the descriptor pool from
which a descriptor is fetched is known when the descriptor's dma address
is being unmapped to its virtual address.  This works fine if all the
descriptors, in particular tx descriptors, are taken from a single pool
and the pool handle is stored as the caller's global.  However if the
(tx) descriptors are created from different pools,callers trying to
unmap a descriptor's dma address would not know from which pool the
descriptor is created until after the unmap is done.

This patch adds a new API to unmap a descriptor's dma address
without any pool handle passed in as a parameter.  This implementation
essentially breaks up the existing knav_pool_desc_dma_to_virt() into two.
First part knav_desc_dma_to_virt() is to loop over the regions to find
out the pool from which the descriptor's dma address is taken and return
the corresponding virtual address.  After caller finished processing the
descriptor with the descriptor's virtual address, it should call the
second part knav_dma_desc_unmap() to complete the unmap.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>