soc: ti: saving qmss clock in driver storage
Some QMSS drivers requires the knowledge of the
clock rate at which the QMSS hardware subsystem
is running at in order to configure other hardware
components correctly.  This patch makes the changes
to save the QMSS clock in its private structure.

This change is backward compatible such that no error
will be posted if no "qmss_clk" is specified.  The
only consequence is that future driver modules
requiring the knowledge of qmss clock rate won't be
supported.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>