hsr/prp: bail out if skb data not pointing to mac header
Currently if hsr/prp driver receives a packet and the mac
header is not pointing to skb->data, then we do an skb_push()
and pass the skb for processing. But there is no check if
the mac header is matching with skb->data. This patch adds the
additional check and bail out.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>