tmp: prueth: add debugfs file to show/write configuration parameters
Add a debugfs interface to configure the configuration parameter MIBs
for HSR/PRP device. This is an intermediate solution to integrate the
net-snmp queries/set function to allow modification of the parameters
through snmp agent. This is expected to be changed in the future to
use proper redundancy device operations. For now this is marked as a
temporary patch for prototyping. While at it, also change the annoying
debug prints when executing the ethtool command by changing it to a
debug print instead of error. Remove the unused shadow variable and
move the error stats display to use debugfs.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>