arm: dts: am57xx-idk-common: add support of hsr/prp vlan pcp to rx queue map"
Usage:

pcp-rxq-map = /bits/ 8 <3 3 2 2 1 1 0 0>;

where (from left to right) low-pcp to high-pcp:

pcp    rxq
-------------
0      3 (q4)
1      3 (q4)
2      2 (q3)
3      2 (q3)
4      1 (q2)
5      1 (q2)
6      0 (q1)
7      0 (q1)

Note 1:
If no pcp-rxq-map is defined, the default is
<3 3 2 2 1 1 0 0>.

Note 2:
If a pcp is mapped to a rx queue (even it is the
default case), it is user's responsibility to make
sure that the rx queue have enough buffers, ie. 48,
to receive at least one regular size frame.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>