prueth: preparatory patch to add run time protocol change support
The run time change of protocol requires moving some of the initialization
for host and port configuration to emac_ndo_open() from probe(). This
requires that some of the internal functions are moved to right place
in the file so that forward declaration of the functions can be avoided.
This also helps in identifying the changes when the above functionality
is added without which, a lot of churn happens in the code and will be
hard to see what is changed. So this preparatory patch is added to help
add support for run time configuration of protocol change in the Ethernet
device. There is no functionality change except that the functions are
moved around.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>