prueth: parse red-rx/red-tx DT property only if eth type is switch
Currently prueth driver parser the red-tx/red-rx DT property always
causing issues on platforms that don't support switch. Fix this by
adding a conditional check to see if the device type is switch and
then parse the checking if the driver is working as switch and parse
the red-rx/red-tx DT property.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>