hsr/prp: improvements to lre statistics handling
Add a statistics for number of supervisor frames send by the driver
so that better accounting can be done for number of supervisor frames
vs others. While at it also fix the cnt_tx_c,cnt_rx_c statistics
counters. As per IEC62439 specification, TXC is number of frames sent
to application interface and RXC is number of frames received at the
interface. Present code does it in the reverse way. So fix the
same by interchanging the variable displayed as part of lre statistics.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>