net: prueth: change rx packet processing to use irq thread instead of NAPI
NAPI requires typically more buffers (some example shows about 10 buffers
of MTU size) to allow enough buffering so that packets get queued for NAPI
processing without receive overflow. Since there is only about 4 buffers
of MTU size in the receive direction, the NAPI processing does introduce
packet loss with MTU size packets. One way to fix this is to use irq thread
to process packets. Test has shown that with this change, there is no
packet loss when MTU sized packets are receieved at a rate as high as
88Mbps. So use irq thread to handle received packets instead of NAPI.

This patch uses netif_rx() instead of netif_receive_skb() as the later is
not expected to be used inside a softirq context. Gets a warning trace
when netif_receive_skb() used. At the transmit side, firmware currently
doesn't support tx interrupts. So the same code to request/free irq is
not done for switch based firmwares (Switch/HSR/PRP).

Note that this breaks compatibility with older firmware as previous version
of the firmware and driver uses a shared interrupt at the receive side.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>