net: prueth: remove unnecessary logs and code comment
The pcp-rxq-map is now not configurable and is static. First 4 levels
of PCP are mapped to first queue and others to the second queue. So
remove the associated log which was applicable when the same is
configurable and the variable, pcp-rxq-map that is unused. Also there
is an unnecessary logs that appear when a link is disconnected while
traffic is ongoing on the wire. This also seems to introduce packet
drop when link fail over happens. So remove the same. Additionally
remove a TODO comment which is not relevant any more.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>