hsr/prp: fix duplicate discard mode handling for PRP
PRP has a duplicate discard mode that can be set to either
duplicate discard by default (0x20) or duplicate accept (0x21).
This causes the mode to be reflected in the SV frame TLV type value
to be either of these values. Currently this can't be changed. So add
a debugfs interface to change this at run time. Mark this as a HACK
since proper way to do this seems to be through the netlink socket
interface for which driver requires more enhancement.

While at it, fix a cut-n-paste error in comments for hsr_mode debugfs
code and also add range check for hsr mode parameter value when received
over debugfs interface.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>