net: ethernet: ti: introduce ptp boundary clock internal sync monitor driver
On some of TI's platforms, more than one timing capable modules are
included. Consider, for example, the CPTS and the two IEP modules on
AM57xx platforms. When these timing modules participate in a PTP
boundary clock (BC) functionality, the PTP clocks implemented by these
modules need to be internally synchronized to the same PTP time of the
PTP domain in which the BC participates. If the internal sync is done
by using the 1PPS generated by the timing modules, it needs to make
sure that only one such module is generating 1PPS at any given
instant of time, for otherwise the SoC life may be damaged over time
when more than one timing module has output signal going against each
other. It is not suffice to rely on the application that implements
that PTP boundary clock software functionality to ensure the one
out-going PPS policy for user may "accidentally" issue PPS enablement
outside of the PTP BC application, for example, through sysfs.

This patch introduces a small driver so that the driver of each BC
participating timing module needs to get permission from this driver
before it can enable its PPS.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>