net:ethernet:ti:cpts: fix double disable issue of 1PPS generation
The timer16 is used to generate the output pulse for both 1PPS and
reference pulse (PEROUT) and should be disabled when both features are
disabled and enabled when either one of them is to be enabled. Both
pps_enable and ref_enable variable are set to -1 to indicate timer16 is
no longer required and should stay that way when the application intends
to disable them again.

This patch adds the required check at cpts_pps_enable() and
cpts_ref_enable()

Signed-off-by: Eric Ruei <e-ruei1@ti.com>