net: prueth: use swork to handle tx timestamp events for RT kernel
This patch fixes the tx timestamp handling by using swork for RT
kernel as it is not approapriate to use schedule_work() to queue
work from hard interrupt context for this kernel. The tx interrupt
handler just schedule a work to handle the tx stamp event instead
of doing everything within the irq handler. So for RT kernel, this
patch uses swork_queue() call and for non RT, it uses schedule_work()
to queue the work for defered processing. Also this addresses the
below BUG: report when debug options are enabled for RT kernel
build.

[  305.880855] BUG: sleeping function called from invalid context
   at kernel/locking/rtmutex.c:987
[  305.880859] in_atomic(): 1, irqs_disabled(): 128, pid: 1080,
   name: ptp4l

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Suggested-by: Grygorii Strashko <grygorii.strashko@ti.com>
Tested-by: Eric Ruei <e-ruei1@ti.com>