prueth: use u16 for firmware lock
As we have two PRUs sharing the same lock it is possible that the both
can obtain the lock and one PRU returns it, while another PRU still
working on the node table. If the ARM core is waiting for the lock, it
may start updating the node table, while the second PRU has not finished
its job.

To eliminate this we allocate u16 for fw_lock, where each PRU will use
one byte of it. So, in this situation if one PRU returns the lock, it
will be still locked by the second.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>