hsr: save tx_flags and tag info of tagged skb
After a tag is created for an egress skb, the tag info and slave port
info is saved in the skb's shared info area so that the saved info can
be returned to user space application by upper core layer.

In this patch, only the support of saving HSR tag info of HSR PTP
over L2 skb is added.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>