prueth: Fix random RT Linux crash seen when running oc/bc application
This patch fixes the random kernel crash seen when running oc/bc
application on an RT Linux kernel. The issue seems to be happening
because in the receive packet processing path, the kernel code
encounters an skb that is already freed. This is fixed by properly
reference counting the skb from xmit handler (through a call to
skb_get() before passing ownership to the ts even handler worker.
As skb is already reference counted, it is freed at the end of
xmit handler through a call to dev_kfree_skb_any() as done with
frames not requiring time stamping.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>
Suggested-by: Grygorii Strashko <grygorii.strashko@ti.com>
Tested-by: Eric Ruei <e-ruei1@ti.com>