prueth: Fix the high CPU usage issue on AM3/4 devices
As soon as prueth Ethernet interface is brought up on an AM3/4
device, the ARM CPU utilization goes to 95% and the console
becomes unresponsive. This is root caused to the IEP Global
configuration register not being initialized to 0x551 which
was the case prior to adding PTP driver. On AM3xxx devices,
PRU Dual EMAC firmware currently uses the IEP counter for FIFO
fill level logic and expect the value to be 0x551. AM4xxx
devices re-use the AM3xxx firmware and has the same requirement.
So fix this by restoring the code to initialize IEP_GLOBAL_CFG_REG
register to 0x511 when PTP is not used.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>