prueth: move mc filter table to DRAM1 from SRAM
Firmware required some changes to the memory map to make room for
VLAN filter table in SRAM. For this, Firmware moved the MC filter table
from SRAM to DRAM1 and this patch updates the Linux code to work with the
firmware with these changes. Firmware also introduced a statistics counter
to count the dropped multicast Ethernet frames at PRU when a multicast
address hash is not set in the filter table.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>