prueth: add support for vlan filter at PRU host receive port
prueth firmware for HSR/PRP implements a VLAN filter table of 512
bytes where each bit in the byte represents a VLAN ID. This allows
filtering out Ethernet frames at the PRU host receive port which
are not part of the VLAN group to which the host has subscribed to.
When a VLAN interface is created, the VID information gets passed
to the driver through ndo_vlan_rx_add_vid() function call.
Similarly driver gets notified of the VID to be removed from the
filter when a VLAN interface is removed through ndo_vlan_rx_kill_vid().
The prueth driver inserts or deletes a VID in the VLAN filter table
as part of these function calls.

Untagged frames are dropped and is not received at the Host when
vlan filter is enabled. This gets enabled when the first VLAN
interface is created. However priority tagged frames gets forwarded
to Host when filter is enabled. These are frames with VID 0.

The firmware also provides a statistics counter to keep track of
the VLAN frames dropped at PRU host receive port. This statistics
counter can be viewed using ethtool command. Finally add a debugfs
file to view the contents of the vlan filter table in memory.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>