prueth: enable untagged vlan frames to workaround node table issue
Currently if vlan filtering is enabled, PRU stops sending supervision (SV)
frames to host since VLAN filtering filter out the untagged SV frames.
As usage of VLAN tag with SV frame is not yet supported, and the
requirement needs clarity, for now enable untagged VLAN frames to Host
by configuring the same in PRU. Without this, node table functionality
is broken. So add this workaround to resolve the issue for now.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>